Large commit message below:
Extending support for other forcefield itp files uncovered a bug where undefined, random failure to resolve a dihedral containing the new bond was observed during polymer extension. The cause of this bug was due to the Bond.from_atoms static method, which would find all bonds between a pair of atoms, and if there was more than 1 (in the case of a duplicated bond between two joining atoms during extension) it would return the second bond. This method is envoked to find bonds which will be assigned angles (and thus dihedrals) when a leaving 'residue' atom is replaced by a new 'monomer' atom. As a result, duplicate bonds were found to share half of the angles each, which caused issues when attempting to resolve dihedrals. This was corrected by ensuring the method returns the bond with angles associated if there are duplicates. However, after polymer extension is complete, the bonds are 'deduplicated' to remove a duplicate, and the duplicate that is removed would be the second one iterated through in Atom 'deduplicate_bonds()' -> thus if the second of the two bonds was the one with the angles it would be lost and also result in issues resolving dihedrals in the next polymer extension step. I believe this bug only surfaced now as the test case I was using had a two-atom repeating backbone, so the angles of the bond created in the previous extension step were needed to resolve the dihedrals in the current extension step. I suspect it may also be attributed to changes in dihedral ordering from introducing support to other forcefields, but I am not sure.

Changes made: deduplicate_bonds() uses Bond.from_atoms() to select the correct, 'empty' bond to delete
              Bond.from_atoms() has an optional argument used to specify if it should return the empty bond or the bond with angles, returns the bond with angles by default